ENH add argNorm version to tsv output to the top of tables as comments
Tsv outputs will now have the argNorm version on the top as a comment, such as: "# argNorm version: 0.6.0".

Tests in test_normalizers.py now skip the first row of the golden files that contain the argNorm version. `get_golden_file` function was added to test_normalizers.py to make changing the procurement of golden files easier.