BUG: account for genes being mapped directly to drug classes without intermediate drugs in `confers_resistance_to()`.
Whenever genes were mapped to drug classes (i.e. immediate child of antibiotic molecule), `confers_resistance_to()` would store these mappings in a temporary list (`backup_drugs`) and only use them if the ARG wasn't being mapped to any other drug.

This strategy fails if an ARG is mapped to a drug class and a drug that falls under another drug class (not the same as the one being directly mapped to the ARG). In this case, the ARG is only mapped to the drug and the mapped drug class information is lost.

Now, `confers_resistance_to()` does not utilize `backup_drugs`. Rather, it maps ARGs to every possible drug and drug class. It then iterates over these drugs and drug classes to check if any of the drugs are children nodes of mapped drug classes. Mapped drug classes that are the parents of mapped drugs are removed (they will be restored by `drugs_to_drug_classes()`).