ENH: relax constraints on matching NCBI genes
Previously, both the gene name and the reference accession were used to locate the mapping for a particular NCBI gene. However, depending on whether or not the output is hamronized (and depending on the ARG annotation tool; Abricate vs AMRFinderPlus), the gene name can be recorded in different ways (e.g., 'NimIJ family 5-nitroimidazole reductase' vs '5-nitroimidazole_reductase_NimJ'). To ensure the mappings of the same genes from hamronized and raw outputs are the same, only the reference accessions are used for the mapping now as they are common for all outputs and can accurately locate mappings.